Remove associated multiplexes when deleting sources.
Fixes #10156

+++ mythtv/libs/libmythtv/sourceutil.cpp
@@ -492,6 +492,17 @@ bool SourceUtil::DeleteSource(uint sourceid)
         return false;
     }
 
+    // Delete the multiplexes associated with the source
+    query.prepare("DELETE FROM dtv_multiplex "
+                  "WHERE sourceid = :SOURCEID");
+    query.bindValue(":SOURCEID", sourceid);
+
+    if (!query.exec() || !query.isActive())
+    {
+        MythDB::DBError("Deleting Multiplexes", query);
+        return false;
+    }
+
     // Delete the inputs associated with the source
     query.prepare("DELETE FROM cardinput "
                   "WHERE sourceid = :SOURCEID");
